+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixes
 - Arch
-  - not properly checking all provided packages by the system when checking for conflicts (upgrade requirements)
+  - not properly checking all provided packages by the system when checking for conflicts (upgrade requirements). it could lead to an unbootable system or package losses.
 - UI
   - some windows not properly being centralized (regression introduced in **0.10.4**)
   - some windows being displayed in the wrong place (regression introduced in **0.10.4**)